# feat: Scene Creation Modify

+++ crates/pi_scene_context/src/scene/command.rs
@@ -8,15 +8,15 @@ use super::ScenePassRenderCfg;
 
 pub struct OpsSceneCreation(pub(crate) Entity, pub(crate) ScenePassRenderCfg, pub(crate) SceneBoundingPool);
 impl OpsSceneCreation {
-    pub fn ops(scene: Entity, passes_cfg: ScenePassRenderCfg, cullingmode: u8, param: [usize;9]) -> Self {
+    pub fn ops(scene: Entity, passes_cfg: ScenePassRenderCfg, cullingmode: u8, param: [i32;9]) -> Self {
         let pool = match cullingmode {
             2 => {
                 SceneBoundingPool::create_oct(
                     (param[0] as Number, param[1] as Number, param[2] as Number),
                     (param[3] as Number, param[4] as Number, param[5] as Number),
-                    param[6],
-                    param[7],
-                    param[8]
+                    param[6] as usize,
+                    param[7] as usize,
+                    param[8] as usize
                 )
             },
             _ => {
